+++ app/src/main/java/com/echoeyecodes/clipclip/models/VideoCanvasModel.kt
@@ -1,5 +1,5 @@
 package com.echoeyecodes.clipclip.models
 
-data class VideoCanvasModel(val width: Double, val height: Double){
+data class VideoCanvasModel(val width: Float, val height: Float){
     var isSelected = false
 }
